+++ packages/common/web/src/Base/createBaseComponent.jsx
@@ -35,6 +35,7 @@ const createBaseComponent = ({
   getEditorBounds,
   onOverlayClick,
   onAtomicBlockFocus,
+  disableRightClick,
 }) => {
   class WrappedComponent extends Component {
     static displayName = createHocName('BaseComponent', PluginComponent);
@@ -243,6 +244,8 @@ const createBaseComponent = ({
       pubsub.set(batchUpdates);
     }
 
+    handleContextMenu = e => disableRightClick && e.preventDefault();
+
     render = () => {
       const { blockProps, className, selection, onDragStart } = this.props;
       const { componentData, readOnly } = this.state;
@@ -332,6 +335,7 @@ const createBaseComponent = ({
           className={ContainerClassNames}
           data-focus={isActive}
           onDragStart={onDragStart}
+          onContextMenu={this.handleContextMenu}
           {...decorationProps}
         >
           {!isNil(link) ? (