+++ packages/plugin-gallery/web/src/gallery-viewer.jsx
@@ -176,6 +176,8 @@ class GalleryViewer extends React.Component {
     </Fragment>
   );
 
+  handleContextMenu = e => this.context.disableRightClick && e.preventDefault();
+
   render() {
     this.styles = this.styles || mergeStyles({ styles, theme: this.context.theme });
     const { scrollingElement, ...settings } = this.props.settings;
@@ -188,6 +190,8 @@ class GalleryViewer extends React.Component {
         ref={elem => (this.container = elem)}
         className={this.styles.gallery_container}
         data-hook="galleryViewer"
+        role="none"
+        onContextMenu={this.handleContextMenu}
       >
         <ProGallery
           items={items}