+++ packages/plugin-image/web/src/createImagePlugin.js
@@ -57,6 +57,7 @@ const createImagePlugin = (config = {}) => {
     uiSettings,
     t,
     isMobile,
+    disableRightClick: config?.uiSettings?.disableRightClick,
     ...rest,
   });
 };